Add files via upload
Briefly summarize The Gaming Room client and their software requirements. Who was the client? What type of software did they want you to design? The Gaming Room client was a mobile game app that was developed for the Android platform. This game was designed to work across multiple devices and platforms.

What did you do particularly well in developing this documentation? I think that the requirements is and the process is key as well as the organization of the layout. This ensures the client is able to understand all the details properly.

What about the process of working through a design document did you find helpful when developing the code? When developing classes and relationships between structures to represent the app components. The domain model as a UML proved to show beneficial understanding the relationship  and its structures.

If you could choose one part of your work on these documents to revise, what would you pick? How would you improve it? I would say the evaluation section. I think I could have further elaborated on the different mobile platforms along side the pros and cons of them all.
How did you interpret the user’s needs and implement them into your software design? Why is it so important to consider the user’s needs when designing? Well the user experience is important because it aims to provide and fulfill all of the user's requests. This will keep the user loyal to future work needed. Providing a meaningful experience is a plus as well for the brand.

How did you approach designing software? What techniques or strategies would you use in the future to analyze and design a similar software application? One strategy I would focus on is discipline which helps decide how and why to strategically design interplay between the design and the business idea.